+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog is loose. Versions are not semantic, they are incremental. Splint
 
 - Update parser to add `:splint/origin-ns` if a symbol's namespace is an alias.
 - `lint/redundant-call`: Make sure the function isn't aliased from another namespace.
+- `lint/defmethod-names`: Skip methods with unquoted dispatch vals. (For example, `(defmethod foo ~some-val ...)`.)
 
 ## 1.22.0 - 2025-11-17